[wip] integrate all sets and custom if any
